Remove methods added to public pooled allocator interfaces
Due to backwards compatibility, we can only add methods to concrete classes.
