+++ buffer/src/main/java/io/netty/buffer/ByteBufAllocatorMetric.java
@@ -25,20 +25,4 @@ public interface ByteBufAllocatorMetric {
      * Returns the number of bytes of direct memory used by a {@link ByteBufAllocator} or {@code -1} if unknown.
      */
     long usedDirectMemory();
-
-    /**
-     * Returns the number of bytes of heap memory that is currently pinned to heap buffers allocated by a
-     * {@link ByteBufAllocator}, or {@code -1} if unknown.
-     * A buffer can pin more memory than its {@linkplain ByteBuf#capacity() capacity} might indicate,
-     * due to implementation details of the allocator.
-     */
-    long pinnedHeapMemory();
-
-    /**
-     * Returns the number of bytes of direct memory that is currently pinned to direct buffers allocated by a
-     * {@link ByteBufAllocator}, or {@code -1} if unknown.
-     * A buffer can pin more memory than its {@linkplain ByteBuf#capacity() capacity} might indicate,
-     * due to implementation details of the allocator.
-     */
-    long pinnedDirectMemory();
 }
