+++ buffer/src/main/java/io/netty/buffer/PoolArenaMetric.java
@@ -152,10 +152,4 @@ public interface PoolArenaMetric extends SizeClassesMetric {
      * Return the number of active bytes that are currently allocated by the arena.
      */
     long numActiveBytes();
-
-    /**
-     * Return the number of bytes that are currently pinned to buffer instances, by the arena. The pinned memory is not
-     * accessible for use by any other allocation, until the buffers using have all been released.
-     */
-    long numPinnedBytes();
 }
