+++ buffer/src/main/java/io/netty/buffer/PoolChunk.java
@@ -588,7 +588,6 @@ final class PoolChunk<T> implements PoolChunkMetric {
         }
     }
 
-    @Override
     public int pinnedBytes() {
         synchronized (arena) {
             return pinnedBytes;
